+++ na-client/js/files/extract-track.js
@@ -3,7 +3,7 @@ import crypto from 'crypto';
 
 import uuid from 'node-uuid';
 
-export default function(musicDir) {
+export default function extractTrack(musicDir) {
     return function(file) {
         const [genre, artist, album, songFile] = path.relative(musicDir, file).split(path.sep);
         const id = crypto.createHash('sha256').update(file).digest('hex');
@@ -19,12 +19,4 @@ export default function(musicDir) {
             id
         };
     }
-}
-
-export function buildFileInfoForBackend(map, track) {
-    map[track.genre] = map[track.genre] || {};
-    map[track.genre][track.artist] = map[track.genre][track.artist] || {};
-    map[track.genre][track.artist][track.album] = map[track.genre][track.artist][track.album] || {};
-    map[track.genre][track.artist][track.album][track.number] = {id: track.id, title: track.title};
-    return map;
 }
