+++ na-client/js/files/file-watcher.js
@@ -1,29 +1,11 @@
 import fs from 'fs';
 import Rx from 'rx';
 
-import extractTrack, { buildFileInfoForBackend } from './files-parsing';
-import { reduceAndMemoize } from '../lib/helper';
+import extractTrack from './extract-track';
+import buildFileInfoForBackend from './build-file-info';
+import reduceAndMemoize from '../lib/reduce-and-memoize';
 
-const CLIENT_KEYFILE = '.clientkey';
-const DS_STORE = '.DS_Store';
-
-export function fileProcessor(filesById, musicDir, files) {
-    const tracks = files.filter(file => {
-        if (file.indexOf(CLIENT_KEYFILE) !== -1) {
-            return false;
-        }
-        if (file.indexOf(DS_STORE) !== -1) {
-            return false;
-        }
-        return true;
-    })
-    .map(extractTrack(musicDir));
-
-    const processTracks = reduceAndMemoize(filesById, 'id', 'file');
-    return processTracks(tracks, buildFileInfoForBackend, {});
-}
-
-export function fileWatcher(monitor, sendFileData, musicDir) {
+export default function fileWatcher(monitor, filesById, sendFileData, musicDir) {
         Rx.Observable.fromEvent(monitor, 'created')
             .filter((file) => {
                 return fs.lstatSync(file).isFile();
