+++ na-client/js/lib/reduce-and-memoize.js
@@ -6,9 +6,3 @@ export function reduceAndMemoize(memo, memoKey, memoValue) {
         }, initial);
     };
 }
-
-export function createWriteStream(stream) {
-    return function(hostedFileData) {
-        stream.write(hostedFileData);
-    };
-}
